Add a ServiceResourceTransformer to let Jersey work properly
When assembled into a shaded jar Jersey would complain about not knowing about
the application/tar mime type.

The problem turned out to be related to how the jersey jars were being included.

The fix was to use the ServiceResourceTransformer in the configuration of the
maven-shade-plugin.  See: http://stackoverflow.com/q/12611213/2336755